requirements.txt
anyio==3.4.0
asgiref==3.4.1
black==21.11b1
click==8.0.3
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pydantic==1.8.2
python-dotenv==0.19.2
PyYAML==6.0
regex==2021.11.10
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.2
typing-extensions==4.0.1
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
